rag-md/pyproject.toml
[project]
name = "rag-md"
version = "0.1.0"
description = "Minimal RAG service over Markdown docs with Qdrant + FastAPI"
authors = [{ name = "Your Name", email = "you@example.com" }]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
  "fastapi>=0.111",
  "uvicorn[standard]>=0.29",
  "pydantic>=2.7",
  "pydantic-settings>=2.2",
  "python-dotenv>=1.0",
  "qdrant-client>=1.9",
  "sentence-transformers>=3.0",
  "torch>=2.2; platform_system != 'Darwin'",
  "torch>=2.2; platform_system == 'Darwin' and platform_machine == 'arm64'",
  "transformers>=4.42",
  "tiktoken>=0.7",
  "markdown-it-py>=3.0",
  "python-frontmatter>=1.1",
  "pyyaml>=6.0",
  "httpx>=0.27",
  "openai>=1.35",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.5",
  "black>=24.4",
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "pytest-cov>=5.0",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
lint.select = ["E", "F", "I", "UP", "B"]
lint.ignore = ["E501"]

[tool.pytest.ini_options]
addopts = "-q -m 'not integration'"
filterwarnings = [
  "ignore::DeprecationWarning",
]
pythonpath = ["."]
markers = [
  "integration: marks tests as integration (deselect with '-m \"not integration\"')",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.build.targets.sdist]
include = ["app", "scripts", "README.md", "pyproject.toml", "tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build" 